scripts/requirements.txt
plexapi
python-dotenv
requests